src/main/persistence/JsonWriter.java
package persistence;

import model.ShoeInventory;
import org.json.JSONObject;

import java.io.File;
import java.io.FileNotFoundException;
import java.io.PrintWriter;

// Represents a writer that writes JSON representation of ShoeInventory to file
public class JsonWriter {
    private static final int TAB = 4;
    private PrintWriter writer;
    private String destination;

    // EFFECTS: constructs a writer to write to the destination json file
    public JsonWriter(String destination) {
        this.destination = destination;
    }

    // MODIFIES: this
    // EFFECTS: opens writer as a new PrintWriter; throws FileNotFoundException if destination
    // file cannot be opened for writing (destination file is not found)
    public void open() throws FileNotFoundException {
        writer = new PrintWriter(new File(destination));
    }

    // MODIFIES: this
    // EFFECTS: writes JSON representation of shoeInventory to a file
    public void write(ShoeInventory shoeInventory) {
        JSONObject json = shoeInventory.toJson();
        saveToFile(json.toString(TAB));
    }

    // MODIFIES: this
    // EFFECTS: closes writer
    public void close() {
        writer.close();
    }

    // MODIFIES: this
    // EFFECTS: adds String to file
    private void saveToFile(String json) {
        writer.print(json);
    }
}
